requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
beautifulsoup4
jinja2
pydantic
pytest